+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [TBD]
+
+  * Added support for Sesame as triplestore synchronization target
+  * Ditched Ruby 1.8 compatibility; iQvoc now requires Ruby 1.9+
+
 ## 4.0.6
 
   * Bugfixes for Triplestore Sync and configuration